feat: update filter all exception
